python/requirements.txt
flask==3.1.3
flask-cors==6.0.2
groq==1.2.0
pymupdf==1.27.2.2
sounddevice==0.5.5
numpy==2.4.4
websocket-client==1.8.0
pyaudiowpatch==0.2.12.8 ; sys_platform == "win32"
